requirements.txt
grpcio
grpcio-tools
protobuf
mypy-protobuf
PyQt6
